Fix template loading
