+++ shortcodes/epfl-fields-of-research-list/view.php
@@ -1,6 +1,5 @@
 <?php
     namespace EPFL\Plugins\Gutenberg\FieldsOfResearchList;
-    $fields = get_query_var('epfl_fields-of-research');
 ?>
 
 <div class="container my-3">
@@ -31,4 +30,16 @@
     </div>
 </div>
 
-<?php load_template(dirname(__FILE__).'/javascript.php'); ?>
+<script type='text/javascript'>
+window.onload = function() {  // wait that jQuery is loaded
+    jQuery(document).ready(function( $ ) {
+        var options = {
+            valueNames: [
+                'field-of-research-name'
+            ]
+        };
+
+        var siteList = new List('fields-of-research-list', options);
+    });
+}
+</script>
